+++ build.rs
@@ -1,6 +1,0 @@
-use anyhow::Result;
-use chksum_build::{setup, BuildScript};
-
-fn main() -> Result<()> {
-    setup(&BuildScript)
-}